requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohttp==3.9.1
aiokafka==0.8.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==25.3.0
azure-core==1.32.0
azure-eventhub==5.11.5
azure-identity==1.15.0
certifi==2025.1.31
cffi==1.17.1; platform_python_implementation != "PyPy"
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6; platform_system == "Windows"
confluent-kafka==2.8.2
cryptography==44.0.2
exceptiongroup==1.2.2; python_version < "3.11"
fastapi==0.104.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kafka-python==2.0.2
msal==1.32.0
msal-extensions==1.3.1
multidict==6.2.0
packaging==24.2
propcache==0.3.0
pycparser==2.22; platform_python_implementation != "PyPy"
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
pyjwt[crypto]==2.10.1
python-dotenv==1.0.1
python-json-logger==2.0.7
redis==5.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.23.2
yarl==1.18.3
